# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * KhetMitra API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DiagnosisSeverity = typeof DiagnosisSeverity[keyof typeof DiagnosisSeverity];


export const DiagnosisSeverity = {
  Low: 'Low',
  Moderate: 'Moderate',
  High: 'High',
  Critical: 'Critical',
} as const;

export type DiagnosisStatus = typeof DiagnosisStatus[keyof typeof DiagnosisStatus];


export const DiagnosisStatus = {
  Pending: 'Pending',
  Diagnosed: 'Diagnosed',
  Treated: 'Treated',
  Recovered: 'Recovered',
} as const;

export interface Diagnosis {
  id: number;
  cropName: string;
  imagePath: string;
  diseaseName: string;
  confidence: number;
  severity: DiagnosisSeverity;
  recommendations: string[];
  /** @nullable */
  affectedArea?: string | null;
  /** @nullable */
  location?: string | null;
  status: DiagnosisStatus;
  createdAt: string;
}

export interface DiagnosisInput {
  cropName: string;
  imagePath: string;
  affectedArea?: string;
  location?: string;
}

export interface Shop {
  id: number;
  name: string;
  address: string;
  latitude: number;
  longitude: number;
  distanceKm: number;
  phone: string;
  products: string[];
  isOpen: boolean;
  rating: number;
}

export type WeatherAlertSeverity = typeof WeatherAlertSeverity[keyof typeof WeatherAlertSeverity];


export const WeatherAlertSeverity = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface WeatherAlert {
  type: string;
  message: string;
  severity: WeatherAlertSeverity;
}

export type WeatherDataDiseaseRisk = typeof WeatherDataDiseaseRisk[keyof typeof WeatherDataDiseaseRisk];


export const WeatherDataDiseaseRisk = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface ForecastDay {
  day: string;
  high: number;
  low: number;
  condition: string;
  rainChance: number;
}

export interface WeatherData {
  location: string;
  temperature: number;
  humidity: number;
  rainChance: number;
  windSpeed: number;
  condition: string;
  diseaseRisk: WeatherDataDiseaseRisk;
  alerts: WeatherAlert[];
  forecast: ForecastDay[];
}

export type SchemeCategory = typeof SchemeCategory[keyof typeof SchemeCategory];


export const SchemeCategory = {
  Financial: 'Financial',
  Insurance: 'Insurance',
  Input: 'Input',
  Technology: 'Technology',
  Other: 'Other',
} as const;

export interface Scheme {
  id: number;
  name: string;
  description: string;
  ministry: string;
  eligibility: string;
  benefit: string;
  applicationUrl: string;
  category: SchemeCategory;
  isActive: boolean;
}

export type DashboardSummaryWeatherRisk = typeof DashboardSummaryWeatherRisk[keyof typeof DashboardSummaryWeatherRisk];


export const DashboardSummaryWeatherRisk = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface MonthlyCount {
  month: string;
  count: number;
}

export interface DashboardSummary {
  totalDiagnoses: number;
  activeDiseases: number;
  nearbyShops: number;
  schemesAvailable: number;
  weatherRisk: DashboardSummaryWeatherRisk;
  recentDiagnoses: Diagnosis[];
  monthlyDiagnosesCount: MonthlyCount[];
}