# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KhetMitra API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnoses:
    get:
      operationId: listDiagnoses
      tags: [diagnoses]
      summary: List all diagnoses
      responses:
        "200":
          description: List of diagnoses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Diagnosis"
    post:
      operationId: createDiagnosis
      tags: [diagnoses]
      summary: Submit a new crop diagnosis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisInput"
      responses:
        "201":
          description: Diagnosis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diagnosis"
  /diagnoses/{id}:
    get:
      operationId: getDiagnosis
      tags: [diagnoses]
      summary: Get a specific diagnosis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Diagnosis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diagnosis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Diagnosis:
      type: object
      required:
        - id
        - cropName
        - imagePath
        - diseaseName
        - confidence
        - severity
        - recommendations
        - status
        - createdAt
      properties:
        id:
          type: number
        cropName:
          type: string
        imagePath:
          type: string
        diseaseName:
          type: string
        confidence:
          type: number
        severity:
          type: string
          enum: [Low, Moderate, High, Critical]
        recommendations:
          type: array
          items:
            type: string
        affectedArea:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
          enum: [Pending, Diagnosed, Treated, Recovered]
        createdAt:
          type: string
    DiagnosisInput:
      type: object
      required:
        - cropName
        - imagePath
      properties:
        cropName:
          type: string
        imagePath:
          type: string
        affectedArea:
          type: string
        location:
          type: string
...