# lib/api-zod/src/generated/types/diagnosis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * KhetMitra API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosisSeverity } from './diagnosisSeverity';
import type { DiagnosisStatus } from './diagnosisStatus';

export interface Diagnosis {
  id: number;
  cropName: string;
  imagePath: string;
  diseaseName: string;
  confidence: number;
  severity: DiagnosisSeverity;
  recommendations: string[];
  /** @nullable */
  affectedArea?: string | null;
  /** @nullable */
  location?: string | null;
  status: DiagnosisStatus;
  createdAt: string;
}