# lib/api-zod/src/generated/types/diagnosisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * KhetMitra API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiagnosisInput {
  cropName: string;
  imagePath: string;
  affectedArea?: string;
  location?: string;
}